Getting ever closer to killing prisma
